requirements.txt
celery==5.4.0
Django==5.0.7
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-filter==24.2
django-redis==5.4.0
django-timezone-field==7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
gunicorn==22.0.0
psycopg2-binary==2.9.9
python-binance==1.0.19
PyYAML==5.3.1
redis==5.0.7
requests==2.25.1
rest-framework-simplejwt==0.0.2
websocket-client
websockets==12.0
python-dotenv